:sparkles: 홈 카테고리 섹션 구현 (#1)
