+++ src/components/domains/home/category/list/styled.ts
@@ -2,7 +2,16 @@ import styled, { css } from "styled-components";
 import { Swiper } from "swiper/react";
 
 export const CustomSwiper = styled(Swiper)`
-  width: 100%;
+  ${({ theme }) => css`
+    ${theme.flex.row("flex-start")};
+    width: 100%;
+
+    .swiper-slide,
+    .swiper-wrapper {
+      width: 72px;
+      height: 108px;
+    }
+  `};
 `;
 
 export const Figure = styled.figure`
